refactor simulator and objective calculator
